Getting sensible numbers!
Still need to check to ensure the outputs indicate what I think they indicate. But eignevalues are now within an order of magnitude of each other, and so anisotropy values are not wholly based on mean intensities anymore. Which is a good sign!
(Repeat due to notes failing to transfer on previous commit.)